PROJ1756: Converted CMSG_ASSERT messages to safe replacements.
ISSU-94

Description:

Replace the existing CMSG_ASSERT macros with something that is "production safe".
These asserts should be used for affirming preconditions and they should
not be activated in practice unless there are some logic errors.
The output will print out a textual representation of the condition that
has failed.
These should be pretty inexpensive as integer comparisons are being made.

Change-Id: Ic98093155f7952f603987d566c7f153d3f6e1556
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>