CR00041903: Lock access to the server socket list.
The list of IPC servers can be accessed from multiple threads, access to the
list needs to be protected. A mutex has been created to protect the list.

The code loop used to process the sockets with available data had an issue
where a socket could be read more than once which could cause a hang. A closed
socket could be reused and because it was not removed from the read_fds could
be processed a second time. The solution is to removed the entry from read_fds
after processing it the first time.

Change-Id: Ib83dfebdd043d96158f8adc42bbede4d0c7066fc
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>