CR00041903: Always peek at TIPC message header.
If an incomplete packet arrives on a TIPC socket the reading process will
block forever, eventually resulting in a healthcheck timeout. While it is
unlikely for an incomplete packet to arrive it is possible therefore to
prevent the process from blocking we should peek the header contents to
check that the full header has arrived.

It is also possible that the message does not arrive at all and again using
the peek will prevent this locking up the switch. The timeouts used are
different the server and client case the server should basically always be
ready however the client needs to wait for a response (RPC) which may take
time to be generated.

Change-Id: I3cc012d994391fd409ddf2074fdfe9cbad183f23
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>