CR-50083: Reduce time of small RPC calls
The introduction of csmg_transport_peek_for_header() creates
a loop which makes sure a received message is complete before
the processing starts. If the message is incomplete this
function makes sure the receive thread doesn't lock up.
This introduced an increase in the timing.

This patch adds a select() call in csmg_transport_peek_for_header()
before the busy loop is entered to improve the response time.

Change-Id: Id63ba82942313211b4c75d9616fdd139d27ce3ff
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>