PROJ1756: Move private information from public header to library private header
This is the first step in separating the CMSG library private and public
interfaces.  This step involves splitting protobuf-c-cmsg.h so that there
is a new header file with the private definitions.

Unfortunately we still have to include this file every where until the
individual header files are also split so that their private info is removed
from their library interface to the application.

Change-Id: Ie4ef6597a3646091521fe57bf774c9b6f82af145
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>