CR-51087: Make separate copies of messages when using loopback transport
The CMSG loopback transport was simply passing a pointer to the send_msg to the
impl function to use as the recv_msg. This meant that if the impl modified the
recv_msg, the send_msg would also be modified.

Now the message is being packed and unpacked in order to create a separate copy
for the impl to use.

Change-Id: I30d500e908528363f30b4cc3a36ffd39807b2367
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>