PROJ1756: Tidied up "show counter messaging" output
Jira: ISSU-40

Tidied up the show counter output for CMSG.

Prepared counter app_name so it can be displayed as it is and counters for
client and server of the same app can be displayed around when sorted.

Also instead of using 40 for the max app_name length now changed to use the
macro (which has been changed to 64).

Change-Id: Ia7261b4208a99fc36cfe7056259b830183212dea
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>