PROJ1756: Do not use too many counter application names
Jira: ISSU-40

Recent change of increasing the max length of app name in cntrd caused that
the maximum number of applications is reduced from 126 to 52. And as CMSG has
added more counter apps, this limit reaches and the following error occurs:

  Init app. Can't get app registration 'CMSG stk.cli.tipc[2] Client', retVal=6.

So dropped "Client" or "Server" from the app name (and re-adjusted the maximum
app name length to 40 in cntrd lib). And updated counter names accordingly.

Change-Id: Ic2e66551f54606095c7a170817d1449d3ad2bf81
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>