CR-50231: Null the recv_msg pointer in the CMSG_FREE macro
There have been cases where the api of an rpc has been called with the recv
msg not being NULL.
This is usually because the call to the api is done repeatedly inside a loop,
and the recv msg pointer is reused each time without setting it to NULL after
it is freed.
If there is then an error on the api call, and no recv msg is returned, it is
possible that the pointer is still pointing at the old (and already freed)
memory, and therefore the recv msg is processed as new (causing either a crash
or a double free).

This patch takes care of this situation by changing the CMSG_FREE_RECV_MSG
macro to also set the msg pointer to NULL after it is freed.

NOTE: To prevent confusion with other free macros, a later patch will rename
the CMSG_FREE_RECV_MSG macro to CMSG_FREE_AND_NULL_RECV_MSG (and all the
places that uses it).

Change-Id: I97375391fb544d46d336cf697f3a98979055a527
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>