CR-50231: Catch Non-NULL recv msg passed to api
There have been cases where the api of an rpc has been called with the recv
msg not being NULL.
This is usually because the call to the api is done repeatedly inside a loop,
and the recv msg pointer is reused each time without setting it to NULL after
it is freed.
If there is then an error on the api call, and no recv msg is returned, it is
possible that the pointer is still pointing at the old (and already freed)
memory, and therefore the recv msg is processed as new (causing either a crash
or a double free).

Passing a non-NULL recv msg to the api is incorrect usage of the api function.

This patch adds a check to the autogenerated api code to ensure that the recv
msg is indeed NULL. If it isn't NULL, it sets it to NULL and logs a debug
message (as this is a potential memory leak).
The CMSG LOG macros are used to display the debug message, so the cmsg-error.h
header is now included in the api.c

Change-Id: I5244c16d0bc185191e855c4a133a533a732b7eff
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>