PROJ1756: generate CMSG transport's unique id for all CMSG servers
Transport's unique id ("tport_id") is not generated for TCP and UDT servers.
For example, CMSG counter shows "snmp.dyn_sync" service for CMSG server
and "snmp.dyn_sync.udt" service for CMSG client.

Now similar to CMS clients, it is generated in cmsg_server_new().

Also fixed to display IP address and port properly for TCP. Displaying IP
address was wrong as it used a wrong pointer. And port number is stored in
the network byte-order so should be converted to host byte-order to display.

And TIPC instance is now printed in two digits (e.g. 01, 02, ...) so service
names are sorted properly on CFC960 in "show counter messaging".

Change-Id: I4fd19509eabb962197b14560cdf96c377af6e865
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>