PROJ1756: Rename error counters for grouping when sorted
The counter names are currently displayed sorted in alphabetical order. With
the current error counter names, this has the side effect of spreading the
error counters amongst the other counter names, as "error" is at the end of
the name.

This patch renames the error counters to move the word "error" towards the
front of the name, so the sorting will group the error counters together.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>