PROJ1756: Fix merge error in autogenerator
An earlier patch accidently changed the name of a dictionary variable to one
that was already in use (suspected merge error).
This resulted in the wrong name being output in the autogenerated service
descriptor.

Even though this field is not currently used for anything, it is still wrong
so this patch corrects it.

Change-Id: I9a72e13ef81ff2daaa3ca223ba75bb95745b752f
Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>