CR-50380: Ensure the socket gets cleared if there is no reply from the server
When there was no reply message from the server the socket wasn't getting cleared.
This change ensures that all the information gets read off the socket.

Also if there is a problem allocating memory to complete the socket recv we
need to log it and shut the socket down.  We just don't know how much we have
to read to clear the socket so it is better to shut it down.

Change-Id: Id5d2598c56cc7f7d2f9746804f6d5ab03e0834a1
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>