CR-50342: POEHW memory leak occurs on linecards after CFC failovers
There is a memory leak when a new subscriber entry is added while the same
subscriber entry exists (in case the subscriber didn't call unsubscribe).
For example, this happens when a CFC reboots and POED re-subscribe events
from POEHW (in a LIF).

Now changed cmsg_pub_subscriber_add() to remove existing old entries before
adding new one. Memory leak happened because the new one was discarded
without freeing.
Simply freeing the new entry may be another solution but I think it may be
better to update the order in the subscriber list based on the new subscription.

Change-Id: Icac736732ec6d5ed508598f496ce36d0b9b29c6b
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>