CR-50342: Update the CMSG counter "Server Connect Closed"
Currently "Server Connect Closed" does not count if the peer disappears without
shutting down the socket properly. This error case is counted in "Server Errors: Recv".

Having "Server Connect Accepts" and "Server Connect Closed" gives an indication
of that the difference is the number of sockets currently connected. But it's not
true because "Server Errors: Recv" also indicates socket closure (and maybe other
cases as well).

So to make it easier to check the server sockets connected, now "Server Connect Closed"
counts socket closure of both orderly shutdown and abnormal disappearance.

Change-Id: I6c6c2045671d27d23e02d326d2ba250cb12020c2
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>