CR-50329: Stop processing CMSG packets without TLV headers
When CMSG received a packet without a CMSG TLV header, it will still attempt to
process it as though it had a CMSG TLV header. This meant that when it received
a packet from an older version of AW+ (5.4.4-0.x), it attempted to read a method
name when none had been sent. In this situation, it would read garbage
characters as the method name.

Note: TLVs were only added to the CMSG header after the 5.4.4-0.0 release.

Now, we only read a CMSG TLV header if we have received enough data for there
to be a TLV header.

Change-Id: If13e0592f43b245ed58396e681c9e39a1787459f
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>