CR-50551: Fix null crash on empty message after unpack
The problem here occurred because two switches running different software were
cabled up to form a stack. One switch was running recent software and the
other was running software from 2 months ago (both mainline daily builds).
In between these two builds a project was delivered to mainline that redefined
a message by re-indexing all the fields in it.

This drastic change in the message definition meant that the server couldn't
unpack the message because it was clear that it's definition of the message was
quite different to the one it received.
If a message fails to unpack on the server side, it returns a NULL pointer up
to the server processing.
Instead of stopping there, the server proceeded to try and use the message
anyway, which results in a NULL pointer dereference inside the server invoke
(and the subsequent crash).

Note: Normally stackd will detect that two software releases are incompatible
and force a software sync to occur. In this case the message that changed (and
caused the problem) was part of the stack discovery process, so the software
compatibility check hadn't been reached yet.

The fix is to bail out once the server can't do any more due to the NULL
message.

Change-Id: I84a66843242ba32ceab66487e5620a22e7c95087
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>