PROJ1756: Pull changes from protobuf-c upstream
Pulled in protobuf-c 1.0.0 from github.

This creates directories for protoc-c and protobuf-c which are
parallel to our cmsg and protoc-cmsg. This makes pulling in changes
from protobuf-c easier. But merging in changes has still to be done manually.
None of our files are updated yet with this.

Pulled in the new protobuf-c make system.

Change-Id: Id6cb60f38885363042a2fac2206a399652e19ec3
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>