LUG-178: Add connection open message to support new TIPC implied connection
In Linux commit:5d21cb70db0122507cd18f58b4a9112583c1e075 the stream sockets
were modified to support implied connection.  This means that if an app attempts
to connect to a port that has not been created (there is no server) the connect
call returns success (0).  On 3.6.11 it returns ECONNREFUSED.

To ensure that our CMSG code keeps operating with our original expectations
this patch edits CMSG TIPC transports to send a CONN_OPEN msg after connecting
to see if the connection has been made.  If the send fails there is no actual
connection and therefore the connect function fails.

Change-Id: I81fa8b6d5b6381323bc867dcce401565affd5c82
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Tim Beale <tim.beale@alliedtelesis.co.nz>