CR-50744: CMSG fix queue_process_mutex to avoid deadlock
This fix is also needed to fix EXFX deadlock issue.

In cmsg_client_queue_process_all(), "client->queue_process_mutex" is locked
while processing the messages in the queue. This means any thread trying to
add a message to the queue will be blocked. However the queue is already
protected by "client->queue_mutex" using "queue_process_mutex" this way is wrong.

This causes a deadlock in EXFX when "exfx_fdb_eventQueuesTask" thread is blocked
due to CPG congestion and "AIS Linstener" thread tries to put a message to
the client queue.

As "queue_process_mutex" is added to protect "client->queue_process_count" variable,
the lock should be put before modifying/accessing the variable (not during
processing the messages).

Change-Id: Ibb47bd68e0f274aed4bbb3745b9d7bf2f9bffb64
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>