MAINT: Add CMSG_MSG_ARRAY_ALLOC()/_FREE() helper macros
This patch adds a helper macro to allocate an array of message structs and
set them up ready for use. Typically code using CMSG_SET_FIELD_REPEATED()
mallocs 2 arrays - an array of pointers and an array of message structs.
The impl/API code doesn't really use the array of structs very much - it's
just needed as somewhere to store the actual message data.

I've reviewed a couple of patches recently that tried to malloc both arrays
in a single piece of memory. This can be required if the message involves
nesting within the message structs - the memory block for the message
structs may need to be allocated in another function, but there's no
reference to the array-block stored in the message (for freeing it later).

Allocating both arrays in a single block of memory adds quite a bit of
complexity to the CMSG impl/API code, and it's not immediately obvious why
it's being done that way. The idea is by moving the code into a common
helper macro/function, it'll hide some of the complexity and the rationale
will be documented in a common place.

Also added corresponding FREE() macro so this code can use the Glib memory
recording, same as what received CMSG messages currently use.

Change-Id: I9386e19e23b54e2ee9e1ea13bc3dcea0bbc68774
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>