CR-52421: Fixed unable to bind ipv6 cmsg transport tcp socket.
The issue was observed when ATMF gui schema was sent from an
ATMF Master to Controller device using Ipv6 cmsg tcp transport.

The issue is a race condition with IPv6 DAD. Until DAD can confirm
that there is no other host with the same address, the address is
considered to be "tentative". While it is in this state, attempts
to bind() to the address fail with EADDRNOTAVAIL, as if the address
doesn't exist.

This patch adds a new flag in cmsg_transport to allow user to
configure TCP socket option IP_FREEBIND for IPv6 which allows binding
to a non-local or non-existent address.

This flag is not set for other transports for now.

Change-Id: I0553163b6314c4dffbb5de115a50c151ed4c4328
Reviewed-by: Hermin Anggawijaya (Angga) <hermin.anggawijaya@alliedtelesis.co.nz>
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Geoff McCaughan <geoff.mccaughan@alliedtelesis.co.nz>