AMFGUI-7: Add ability for the impl to keep the recv msg memory
For applications that use CMSG structures natively, being able to keep a
message received from CMSG without copying it first is really useful.

Currently, only messages received from an API call (client side) are owned
by the application.

This patch allows an application to take ownership of messages received at the
server, so the server won't delete them after calling the IMPL.

There are two options for using this:
Call cmsg_server_app_owns_current_msg_set () on a per message basis from
inside the IMPL. This flag will be automatically be reset after the IMPL exits,
and is useful for situations where you only want to keep some messages.

Call cmsg_server_app_owns_all_msgs_set (TRUE/FALSE) on a global basis. This flag wont
be automatically reset so is useful when you always want to keep the received
messages.

WARNING: Ownership also means responsibility for freeing the messages.
WARNING: There is no way to change ownership of an existing message once the IMPL
function exits. Whatever the flag was set to when the message is received
determines the owner for the rest of the life of the message.

Change-Id: Iac690604bff43a6c6072216b42aac945443f92bc
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>