AMFGUI-7: Add const restriction on msg in serverSend function
Currently messages received via an IMPL are specified as const due to the IMPL
function signature.
Now that CMSG allows applications to keep messages received in an IMPL the
const limitation applies to those kept messages.
This makes it hard for applications to re-send any received messages because
the serverSend function doesn't have the const requirement for the passed in
message.

This patch makes the message parameter to the autogenerated serverSend
function const, as CMSG doesn't change the message anyway, and adding it
allows for const messages to be sent.

Note: the complete way to solve this would be to remove the const limitation
from the IMPL signature. However to do that will require every current IMPL to
have its signature changed, so needs more consideration as to whether it is
worth it.

Change-Id: I51ca50431a31d9a25c58a14ad23ae78f7b0d71f2
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>