MAINT: Fix transport ID for IPv6 TCP transport.
A previous patch adding IPv6 support to the TCP transport missed a case
where the transport identifier code would always use the position of the
IPv4 address when printing it to a string.  This caused IPv6 addresses
to be somewhat garbled.  This has been fixed by moving printing of the
ip address into the if that determines the port to print, additionally
an ipv6 address is enclosed in square brackets as is done in URLs to
disambiguate the : used as a port separator.

Additionally the CMSG_MAX_TPORT_ID_LEN define was set at 32, this is not
long enough to allow for a full length IPv6 address as well as the other
information present in the TCP transport ID.  This has been increased to
64.  Although the tag is included into a counter name it is included after
some other names and could already be truncated at a length of 32 so
increasing it's length for the sake of better output in logs and other
debug is reasonable.

Change-Id: I61b2effc4c83c84277b9a4930ce90ab4a542df75
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>