FIBMON-86: Fix client receive logic for TCP type connections
Up until now the only places using TCP type connections with CMSG
have been 'One-way' transport. Therefore the issue with the client
receive logic for TCP has only now been noticed due to fiber monitoring
requiring a 'RPC' type transport.

Specifically, the code that calculated whether there was data following
the CMSG header was incorrect. The code needs to check whether the
dynamic length of the message is 0, rather than the message length
in the header being 0.

There are also a couple of other places in the TCP receive code that
look incorrect to me. Since TCP type connections are barely used in AW+
I figure we may as well just use the same logic that is implemented
for TIPC type connections as these are used everywhere and the logic
has obviously been well tested. Since at this level both the TIPC and
TCP receive code is just operating on bytes in a socket (which should
just be a CMSG packet) it is safe in my opinion to use the same logic.

Change-Id: I8cedfb338b6e69889b0664684fa0fdc9b9a316b3
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>