CR-54011: Add mechanism to configure CMSG logs to be debug instead of error level
There was a patch introduced for CR-53741 that filtered out CMSG errors from atmf_topo process
interactions as they were of no interest to users. As a side effect of that change some log
filters were present in the running configuration which is not acceptable.

After some correspondence with Tim B an alternative approach was suggested to remove ATMF
specific log filtering and instead to modify CMSG to allow suppression of logs (all of which
are reported as errors by default) if required. This is a cleaner approach so this patch
enables the lowering of CMSG log levels from errors to debug during the CMSG client creation.

Corresponding patches are in ATMF to suppress the CMSG log level when creating the client and
in in IPI to remove the ATMF specific log filtering which will resolve this particular issue.

Change-Id: Ia2f50a7c7f6fca26cb83b7113a952cedeb44588f
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>