CR-53933,CR-54039: Avoid CMSG errors from MUX connecting too early
Sometimes clients will try to connect to a remote server too early,
before the server socket has been published. This results in rather
unsightly CMSG debug messages (i.e. "error: Broken pipe").

We can avoid this by waiting until the TIPC port is published. This just
involves the client using the TIPC topology subscription service to find out
when a port is published. However, this adds quite a bit of overhead to the
client to manage its topology subscription (listen on the socket, process
the event, etc). So instead we can add a dynamic query to the topology
service that just returns TRUE/FALSE if a port is currently published.

This is probably not particularly efficient, as it's creating a new topology
subscription socket for every query. However it means the client can now
check the status of the server port with just a single line of code.

Change-Id: I0c691413192d779edc03b72433dc18997c10adcf
Reviewed-by: Geoff Mccaughan <geoff.mccaughan@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>