ENGBL-587: Split out the send and receive of the client invoke functionality
To allow a composite client to call the send and receive parts of a
child client's invoke the existing invoke functionality needs to be
refactored. Specifically, the send and receive parts are now in separate
functions and accessed via function pointers on the client structure.

Standard clients now use a common invoke function:
 - Call the send_invoke function
 - If a recv_invoke function exists, call that next (only 2-way communication
   needs a recv_invoke function).

Eventually a composite client will be able to call the send/recv parts
of the invoke of each child client simply from the function pointers
on the child client structures.

Change-Id: I17137c57e932cd1ca398962012d3e840b68e9d2f
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>