ENGBL-587: Move freeing of unknown fields into the invoke functionality
Currently the unknown fields in a message are freed in the autogenerated
api code. Eventually when a composite client is used an array of messages
will be returned, not a single message. Therefore we need to move the
freeing of unknown fields into the invoke functionality. Now when a
composite client is used each child client will free the unknown fields
before passing the message to the composite client to return via the api.

Change-Id: Ibbe498345fc2886df3e0686fe0950eace4529504
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>