ENGBL-517: Implement composite client
Implement a composite client so that a message can be sent to a group of
clients that respond asynchronously instead of serially. This allows
message processing to occur in parallel on all stack members.

The auto generated code has been modified to collect all the responses.
Existing clients will keep working as they used to and will not need to
be changed.

Change-Id: If878cf052d1669910534fc29ff7812d13c5dad47
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Tim Beale <tim.beale@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>