ENGBL-642: Make cmsg_client_invoke_rpc_send() and cmsg_client_invoke_oneway() near identical
The only difference now is that cmsg_client_invoke_oneway has queueing functionality.

This patch is done purely for the benefit of the next patch that will turn these
two functions into a single function that can be used in both places. Adding in
the profiling debug, moving the counters and adding a call to connect the client
(which previously would have always been connected since we didnt call this) shows
they are identical apart from the queueing.

Change-Id: I097bbc327538ac5596fb20d2cd8eb8449553d9f7
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>