ENGBL-642: Refactor cmsg_client_invoke_oneway and rename to cmsg_client_invoke_rpc_send
These two functions are identical apart from the queueing functionality.
Therefore refactor the queueing functionality to be outside of the function
and use the same function everywhere. It makes sense that a one-way client
should not need its own invoke_send function as in theory it is really just
the same as the send part of a 2-way rpc call.

Change-Id: I538cca75f5ff86561c5deb728f9a440541c1d14c
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>