ENGBL-517: Add NULL pointer checks
On BCM, when setting up the MUX for the local stack node, a CMSG client
is not created (because no IPC is needed for the local node). This ends
up trying to add a NULL client in cmsg_composite_client_add_child which
promptly crashes. Eventually we'd like to have a CMSG client for the
local stack node in BCM (loopback or otherwise), so just add NULL
pointer checks in the CMSG code (rather than changing the MUX code,
which is another option).

Also fix a comment (the variable name changed)

Change-Id: I09ea3a7de3bcbdef2af1876cfeabe4741d753cc1
Reviewed-by: Tim Beale <tim.beale@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>