CR-55304: Pretend we timed out if select() returns EINTR
Even though a user may set SA_RESTART to ensure signals don't interrupt
system calls, due to the timeout value it is implementation dependent
whether the select system call is restarted (with the same timeout) or
simply returns with EINTR. Linux appears to return with EINTR.

Currently if the select call in cmsg_server_receive_poll or
cmsg_server_receive_poll_list is interrupted a log message is printed.
Instead of printing a log message catch the case where EINTR is returned
and pretend that the select call instead timed out.

Another potential option would be to simply call the select function
again. However, since both of these functions are for the user to poll
with it seems to make slightly more sense to pretend the 'poll' found
no data, rather than restart it and potentially take longer than the
timeout the user originally specified.

Change-Id: I58ac66b9996675692e4e6d97021dcb22c7f10df9
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>