CR-55355,ENGBL-642: Only connect before sending messages
The following commit

commit 23518ec ("ENGBL-642: Make cmsg_client_invoke_rpc_send() and cmsg_client_invoke_oneway() near identical")

Changed one-way clients to connect at the start of the invoke function.
This seems to break publishers which return ETIMEDOUT when trying to
connect at this point. Probably the correct thing to have done would
have been to remove the connect from the 2-way case (as
_cmsg_client_buffer_send_retry_once() will call connect anyway.

Therefore remove the cmsg_client_connect() call at the start
of cmsg_client_invoke_send().

Change-Id: I5cf3c8886f9f06b3e4925d0b7bebf629267f088b
Reviewed-by: David Thomson <david.thomson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>