AMFSEC-474: Add support for using UNIX sockets for CMSG
This patch adds support for the use of CMSG UNIX sockets for IPC.
This removes the possibility that the socket communication can be
intercepted by a third party on another device.

This patch is part of a set of 4 patches across 4 repositories (atmfd, ipc,
protobuf-c and ipinfusion) and
need to be delivered together.

Change-Id: I6541edd0e59542ccb89736ed52ab3d90af1dec81
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>