AMFSEC-299: Add support for secure (encrypted) TCP RPC
This patch adds support for encrypted CMSGs using a new transport :-
secure TCP. The application that uses the secure TCP transport is
responsible for encrypting and decrypting the data stream. This is
archived via callback routines that allow the application access to
the CSMG data to encrypt or decrypt the data.

This patch is part of a set of 2 patches across 2 repositories (atmfd and
protobuf-c) and need to be delivered together.

Change-Id: I6e9879db7199a5d56e2f139e7144e85f0304e935
Reviewed-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>
Signed-off-by: Ridge Kennedy <ridge.kennedy@alliedtelesis.co.nz>