AMFSEC-299: Add support for an accept and connect callbacks in crypto mode
Support was added for a second encrypted TCP RPC in ATMF that used a
different calling style to protobuf-c. To support this different style
callbacks were added for the accept and connection events, so that the
client socket descriptor can be passed to the client/server application.
This allows the previously passed server and client context to be replaced
with the clients socket descriptor in calls to encrypt and decrypt traffic.
The aim was to reduce the applications use of protocbuf-c's client and
server data structures.

Change-Id: Ibbc817792155583baaf71b9e6f210466b50765c3
Reviewed-by: Geoff Mccaughan <geoff.mccaughan@alliedtelesis.co.nz>
Signed-off-by: Geoff McCaughan <geoff.mccaughan@alliedtelesis.co.nz>