AMFSEC-549: Fix main line merge issue
Include fix for CR-55737.

CR-55738: Change the way CMSG are read
MSG_WAITALL will cause the recv call to block until all data has been
received, but it appears that in cases where the size of the data to
be received is close to or larger than the size of sockets receive
buffer it may be possible for the TCP connection to deadlock as the
receiver is waiting for the sender to send more, and the sender is
waiting for the receiver to receive more. e.g One of the core files had
143885 bytes of data.

To avoid this situation, the data is now read in chunks that are
significantly smaller than the receive buffer.

This is a investigatory patch, ER-1152: has been created to
create a permanent solution to this problem.

When ATMF topo process was created, We supported only 126 nodes
connected to a controller.  As of 5.4.6-2 we support 306 atmf
nodes and 1000's of Guest nodes. There is potentially too much
information being transferred between Remote area-masters and
the Controller being sent in a single CMSG. This can cause the
atmf_topo process to crash randomly and lost of data between Areas
and the Controller.

Patch provided by Ridge, Tested by Bronwyn

Change-Id: Ib3c4652bb70bd9dc9e4de1d5822dadcf6ab266b6
Reviewed-by: Graeme Campbell <graeme.campbell@alliedtelesis.co.nz>
Signed-off-by: Geoff McCaughan <geoff.mccaughan@alliedtelesis.co.nz>