INTSTAT-36: Tidy up and delete unnecessary files from protobuf-c ATL repo
This repo is used predominantly to build CMSG and related functionality.
Previously we used to also build the protoc-c compiler from here, however
this has been changed to build from the actual upstream protobuf-c package.

We no longer need any of the code related to protobuf-c in this repo (apart
from what is in the cmsg directory), therefore delete the files and tidy up
the Makefile/configure files.

Change-Id: I3892272d14bd83d0961bde8e62fdee11c771c596
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>