INTSTAT-38: Autogenerate cmsg_sub_service.pb-c files at compile time
Rather than manually generate these files and commit them to the repo
simply add a step to the Makefile to generate the files using the
protoc-cmsg compiler.

Technically the protoc-c compiler should be used but there are problems
due to the copied protobuf-c code in the cmsg directory. This will be
fixed in the future.

Change-Id: I3f8f1b8ff1d89502938c06cf5cae6f52d1d7bb45
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>